more renames
using `builder` insted of `build` for the FFI::Build instance so
as not to conflict with the `build` method already existing.  The
verb for `build` (existing method) should probably be different
since we aren't building a c file, we are building the object
file that it generates.  But not sure what a better verb would be.
synthesize?